pyproject.toml
[project]
name = "lavender-ledger"
version = "0.1.0"
description = "Personal expense tracking with PDF bank statement ingestion and Claude-powered categorization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "pdfplumber>=0.11.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
